+++ src/error.ts
@@ -1,7 +1,0 @@
-export class ConfigError extends Error {
-  constructor(message: string) {
-    super(message);
-    this.name = 'ConfigError';
-    this.stack = new Error().stack;
-  }
-}