+++ test/fixtures/fileTypes/js/db.js
@@ -1,0 +1,6 @@
+module.exports = {
+  fileType: 'js',
+  host: 'the db host',
+  username: 'the db username',
+  port: 3306,
+};